day1/radar/requirements.txt
numpy>=1.21.0
rasterio>=1.3.0
matplotlib>=3.5.0
Pillow>=9.0.0
opencv-python>=4.5.0
torch>=2.0.0
torchvision>=0.15.0 
